+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## In progress
 - GitHub/CI: Fixed displaying failed workflow runs on pull requests
   which succeeded afterward
+- GitHub/Important: Improved rendering of Markdown sections
 
 ## v0.1.0, 2025-02-20
 - Started using `GH_TOKEN` environment variable instead of `GITHUB_TOKEN`,